rename: styled component 분리
