+++ fe/src/components/Modal/Modal.tsx
@@ -1,33 +1,10 @@
 import React from 'react';
-import styled from 'styled-components';
 import ModalPortal from 'Portal';
 import Personnel from 'components/Modal/Personnel/Personnel';
 import { useAddSearch, useSearch } from 'context/SearchProvider';
 import Period from 'components/Modal/Period/Period';
 import Price from 'components/SearchBar/Price/Price';
-
-const BackGround = styled.div`
-  position: fixed;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  background: rgba(0, 0, 0, 0.2);
-`;
-
-const ModalBlock = styled.div<{ position: number }>`
-  width: ${({ width }) => width}px;
-  height: max-content;
-  padding: 55px;
-  position: absolute;
-  top: 200px;
-  border-radius: 60px;
-  background: #fff;
-`;
+import { BackGround, ModalBlock } from './Modal.styled';
 
 export default function Modal({
   shown,
