+++ fe/src/components/Modal/Period/Callendar/Callendar.tsx
@@ -1,8 +1,8 @@
 import React, { Dispatch, useState } from 'react';
-import styled from 'styled-components';
 import { ReactComponent as LeftIcon } from 'img/svg/left.svg';
 import { ReactComponent as RightIcon } from 'img/svg/right.svg';
 import { addSearchType } from 'components/SearchBar/types';
+import { StyledCallendar, CallendarTitle, DayOfWeek, PastDay, FutureDay } from './Callendar.styled';
 
 interface CallendarType {
   callendarNum: number;
@@ -10,55 +10,6 @@ interface CallendarType {
   checkOut: number;
   addSearch: Dispatch<addSearchType>;
 }
-
-const CallendarTitle = styled.h1`
-  font-weight: 700;
-  font-size: 16px;
-  margin: 0px 60px;
-  width: max-content;
-`;
-
-const StyledCallendar = styled.table`
-  font-size: 18px;
-  width: 336px;
-  height: 336px;
-  h1 {
-    margin-left: auto;
-    margin-right: auto;
-  }
-  th {
-    vertical-align: middle;
-  }
-  td {
-    font-weight: 400;
-    font-size: 12px;
-    width: 40px;
-    height: 40px;
-    vertical-align: middle;
-    border-radius: 50%;
-  }
-`;
-
-const DayOfWeek = styled.td`
-  color: #828282;
-`;
-
-const PastDay = styled.td`
-  color: #bdbdbd; ;
-`;
-
-const FutureDay = styled.td`
-  &:hover {
-    background-color: #333;
-    cursor: pointer;
-  }
-  ${({ isCheckDate, isPeriod }) => {
-    let color = '';
-    if (isPeriod) color = '#F5F5F7';
-    if (isCheckDate) color = '#333';
-    return `background-color:${color}; `;
-  }}
-`;
 
 const getToday = () => {
   const date = new Date();
