+++ fe/src/components/Modal/Period/Period.tsx
@@ -1,32 +1,7 @@
 import React from 'react';
-import styled from 'styled-components';
-import Callendar from 'components/common/Callendar';
+import Callendar from 'components/Modal/Period/Callendar/Callendar';
 import { ModalProps, PeriodType } from 'components/SearchBar/types';
-
-const StyledCallendarModal = styled.div`
-  display: grid;
-  place-items: center;
-  grid-template-columns: repeat(auto-fit,minmax(50%,auto));
-  ${({ callendarNum }) => {
-    const maxNumOfLine = 2;
-    return (
-      callendarNum > maxNumOfLine &&
-      `
-  overflow-y: scroll;
-  height: 35vh;
-  `
-    );
-  }}
-
-  button {
-  padding: 0;
-  background: transparent;
-  border: none;
-  &:hover {
-    cursor: pointer;
-  }
-
-`;
+import StyledCallendarModal from './Period.styled';
 
 export default function Period({ search, addSearch }: ModalProps<PeriodType>): JSX.Element {
   const checkIn = Number(search.checkIn);
