+++ fe/src/components/SearchBar/Price/Price.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SearchSection from 'components/common/Section';
+import SearchSection from 'components/common/Section/Section';
 import { PriceType, SectionProps } from 'components/SearchBar/types';
 
 export default function Price({ info, search }: SectionProps<PriceType>): JSX.Element {
